OAK-6759: Convert oak-blob-cloud-azure to OSGi R6 annotations (#1112)
* OAK-6759: Convert oak-blob-cloud-azure to OSGi R6 annotations

* OAK-6759: Convert oak-blob-cloud-azure to OSGi R6 annotations - organize imports

* OAK-6759: Convert oak-blob-cloud-azure to OSGi R6 annotations - move statisticsProvider into impl so no annotation inheritence is needed
